% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_deriv}
\alias{calc_deriv}
\title{Calculate derivatives of vector of density data}
\usage{
calc_deriv(
  density,
  percapita = FALSE,
  subset_by = NULL,
  time = NULL,
  time_normalize = NULL
)
}
\arguments{
\item{density}{Density data to calculate derivatives of
Note that data must be sequentially in unique sets and then 
in order when passed to \code{calc_deriv}}

\item{percapita}{When percapita = TRUE, the differences of density are 
divided by density}

\item{subset_by}{if subset_by is provided, it should be a vector (same 
length as density), the unique values of which will 
separate calculations}

\item{time}{Vector of time values provided as a simple numeric
(e.g. number of seconds, rather than a POSIX-based time format)}

\item{time_normalize}{When time_normalize = TRUE, the difference will be 
normalized for the time_normalize value (e.g. if time 
is provided in seconds and the difference per hour 
is wanted, time_normalize should = 3600)}
}
\value{
A vector of the derivative values (or per-capita derivative values)
        the same length as \code{density}, with \code{NA} appended to the end
}
\description{
Provided a vector of density values, this function returns (by default) the
difference between sequential values
}
