% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{from_excel}
\alias{from_excel}
\title{A function that converts base-26 Excel-style letters to numbers}
\usage{
from_excel(x)
}
\arguments{
\item{x}{A vector of numbers in Excel-style base-26 letter format}
}
\value{
A vector of numbers in base-10
}
\description{
A function that converts base-26 Excel-style letters to numbers
}
