\name{growth.curves.pkg-package}
\alias{growth.curves.pkg-package}
\alias{growth.curves.pkg}
\docType{package}
\title{
\packageTitle{growth.curves.pkg}
}
\description{
\packageDescription{growth.curves.pkg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{growth.curves.pkg}
\packageIndices{growth.curves.pkg}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{growth.curves.pkg}

Maintainer: \packageMaintainer{growth.curves.pkg}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
