% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{import_widemeasures}
\alias{import_widemeasures}
\title{Import widemeasures}
\usage{
import_widemeasures(
  files,
  extension = NULL,
  startrow = NULL,
  endrow = NULL,
  startcol = NULL,
  endcol = NULL,
  header = TRUE,
  sheet = NULL,
  wide_names = NULL
)
}
\arguments{
\item{files}{A vector of filepaths (relative to current working directory)
where each one is a widemeasures set of data}

\item{extension}{(optional) the extension of the files,"csv", "xls", or "xlsx"
(will attempt to infer extension if none is provided)}

\item{startrow, endrow, startcol, endcol}{(optional) the rows and columns where
the data is located. If none provided assumes the entire
file is data}

\item{header}{Boolean for whether there is a header to the data. If FALSE
columns are simple numbered. If TRUE is the row above
\code{startrow} (if startrow is specified) or the first row
of the input files (if startrow is not specified)}

\item{sheet}{The sheet of the input files where data is located (if input
files are .xls or .xlsx). If not specified defaults to the first
sheet}

\item{wide_names}{Names to give the widemeasures read in. By default uses the
file names if not specified}
}
\value{
A list of widemeasures named by filename
}
\description{
A function that imports widemeasures in files into the R environment
}
\details{
startrow, endrow, startcol, endcol, timecol, sheet and extension 
can either be vectors or lists the same length as files, 
or a single value that applies for all files
}
