% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{infer_names}
\alias{infer_names}
\title{A function that handles name inference logic}
\usage{
infer_names(
  startrow = NULL,
  endrow = NULL,
  startcol = NULL,
  endcol = NULL,
  infer_colnames = TRUE,
  infer_rownames = TRUE
)
}
\description{
TODO: move logic from read_blocks to here
}
