% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{moving_average}
\alias{moving_average}
\title{Moving average smoothing}
\usage{
moving_average(my_data, window_width, subset_by)
}
\arguments{
\item{my_data}{Vector of numeric data (typically density data)
Data is assumed to be in the correct order to be smoothed
sequentially}

\item{window_width}{Number of data points wide the moving average window is}

\item{subset_by}{Vector of strings or factors. Each unique value of the
\code{subset_by} vector will be smoothed independently}
}
\value{
Vector of smoothed data, with NA's appended at the end
}
\description{
This function uses a moving average to smooth data
}
