% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_blocks}
\alias{read_blocks}
\title{Read blockmeasures}
\usage{
read_blocks(
  files,
  extension = NULL,
  startrow = NULL,
  endrow = NULL,
  startcol = NULL,
  endcol = NULL,
  sheet = NULL,
  metadata = NULL,
  block_names = NULL,
  infer_colnames = TRUE,
  infer_rownames = TRUE,
  wellnames_Excel = TRUE
)
}
\arguments{
\item{files}{A vector of filepaths relative to the current working directory
where each filepath is a single plate read}

\item{extension}{(optional) the extension of the files:
"csv", "xls", or "xlsx"
If none provided, \code{read_blocks} will infer file extension from
provided filenames}

\item{startrow, endrow, startcol, endcol}{(optional) the rows and columns where 
the measures data are in \code{files},
can be a vector or list the same length as \code{files}, or
a single value that applies to all \code{files}
If not provided \code{read_blocks} will try to infer it}

\item{sheet}{(optional) If data is in .xls or .xlsx files, which sheet it 
is located on. Defaults to the first sheet if not specified}

\item{metadata}{(optional) non-spectrophotometric data that should be associated
with each read blockmeasures. A list where each item in the
list is a vector of length 2. Each vector should provide the
row and column where the metadata is located in the blockmeasures
input file. If the list is named those names will be inherited
to the output metadata.}

\item{block_names}{(optional) vector of names corresponding to each plate
in \code{files}. If not provided, block_names are inferred
from the filenames}

\item{infer_colnames, infer_rownames}{Booleans for whether column names
and rownames should be inferred.
If TRUE and startrow or startcol are specified, respectively, 
column or row names will be inferred as the previous column
or row, respectively.
If both \code{infer_colnames} and \code{infer_rownames} are
TRUE and startrow or startcol are not specified,
the first row and first column will be taken as column names
and row names, respectively, if the row 1 column 1 cell is 
empty
If FALSE, rows will be numbered with an "R" prefix and columns
will be numbered with a "C" prefix}

\item{wellnames_Excel}{If row names and column names are not provided in the
input dataframe as specified by \code{infer_colnames}
and \code{infer_rownames}, then names will be generated
automatically.
If \code{wellnames_Excel} is TRUE, generated names
will use Excel-style base-26 lettering for columns
and numbers for rows. 
If \code{wellnames_Excel} is FALSE, rows and columns
will be numbered with "R" and "C" prefixes.}
}
\value{
A list where each entry is a list containing the block measures data
        followed by the block_names (or filenames, if block_names is not 
        provided) and any specified metadata.
}
\description{
A function that reads block measures into the R environment
}
