% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{smooth_data}
\alias{smooth_data}
\title{Smooth data}
\usage{
smooth_data(algorithm = "loess", x, y, formula = NULL, ...)
}
\description{
This function calls other functions to smooth growth curve data
[work in progress]
}
