% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{to_excel}
\alias{to_excel}
\title{A function that converts numbers into base-26 Excel-style letters}
\usage{
to_excel(x)
}
\arguments{
\item{x}{A vector of numbers in base-10}
}
\value{
A vector of letters in Excel-style base-26 format
}
\description{
A function that converts numbers into base-26 Excel-style letters
}
