% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{uninterleave}
\alias{uninterleave}
\title{Uninterleave list}
\usage{
uninterleave(interleaved_list, n)
}
\arguments{
\item{interleaved_list}{A list of R objects}

\item{n}{How many output sub lists there should be (i.e. how many groups
the interleaved list should be divided into)}
}
\value{
A list of lists of R objects
}
\description{
Takes a list that is actually interleaved elements from multiple sources
and uninterleaves them into the separate sources
For instance, a list of blockmeasures that actually corresponds to two
different plates can be split into two lists, each of the blockmeasures
corresponding to a single plate
}
