% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{widen_blocks}
\alias{widen_blocks}
\title{Widen blockmeasures}
\usage{
widen_blocks(
  blockmeasures,
  wellnames_sep = "_",
  nested_metadata = NULL,
  colnames_first = TRUE
)
}
\arguments{
\item{blockmeasures}{Blockmeasures, either a single data.frame or a list of
data.frames}

\item{wellnames_sep}{String to use as separator for well names between 
rowname and column name}

\item{nested_metadata}{A Boolean indicating the existence of nested metadata
in the \code{blockmeasures} list, e.g. as is typically
output by \code{read_blocks}. If NULL, will attempt to
infer existence of nested metadata}
}
\value{
A single widemeasures data.frame
}
\description{
Takes blockmeasures and returns them in a widemeasure format
}
