% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{write_blockdesign}
\alias{write_blockdesign}
\title{Write block designs to csv}
\usage{
write_blockdesign(designs, file, ...)
}
\arguments{
\item{designs}{data.frame or matrix (if one design), or list of data.frames
or matrices, to be written to file}

\item{file}{The filename (if a single design), or vector of file names
(if multiple designs)}
}
\value{
Nothing, but R objects are written to files
}
\description{
This function is basically just a wrapper for write.csv that also works
when handed a list of matrices/dataframes
If \code{designs} is a list, \code{names(designs)} will be placed in the
1,1 cell of each block
}
