% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{merge_tidydesign_tidymeasures}
\alias{merge_tidydesign_tidymeasures}
\title{Merge tidydesign with tidymeasures
(work in progress)}
\usage{
merge_tidydesign_tidymeasures()
}
\description{
Merge tidydesign with tidymeasures
(work in progress)
}
